Timer12 remove, cam config, parameters tuning

+++ Project_Dancer/Dancer/vision.h
@@ -10,8 +10,8 @@
 
 static BSEMAPHORE_DECL(rec_traj_ready_sem, TRUE);
 
-uint16_t get_real_dist_mm(void);
-uint16_t get_hor_dist_mm(void);
+float get_real_dist_mm(void);
+float get_hor_dist_mm(void);
 void process_image_start(void);
 void signal_rec_traj_sem(void);
